Remove rust
I do want to do a project using rust, and I want to do a project using
rust WASM, but I don't think this can be that project. It's too
complicated on its own right, and a lack of enthusasm for a rust project
was curbing my enthusaisam to do this project.

I'm going to spend some time setting this up as a pure javascript with
Typescript JSDocs project. Going to try to go with no frameworks given I
should just need a single page with a canvas on it.