Setup a "new" button
I now have a button to add new smiley faces. Just as a proof of concept
I guess because this probably isn't the model I want to be going with in
the end (I guess I'll have modes, like, a placing mode?).

This also required a lot of css wrangling. And other than an overflow:
hidden that I feel a bit weird about, I think the css is better for it.