Add mouse tracking
This ended up requiring far more thought that expected. Ended up on some
huge detour over software design patterns for the two kinds of vector...

Lots of code for not much impact here.

But the upshot is that the mouse position is now tracked by the state.
It currently logs it when it updates.

I think I may have added the UI for the actual modes I want to add now
as well.