Setup the positions canvas
Items are now drawn to the position canvas which is then drawn (at
offset stored in state) to the display canvas. This is in preparation
for pan and zoom functionality.

I also added recentre and zoom buttons to the controls, though they do
nothing. I think the controls is a bit wide... but that's a problem for
later.