Split state into a canvas and UI controller
This has made the code of both much simpler, and clarified a bunch of
things about the design of the application. Drawing, for instance,
should probably be fully owned by the canvas controller. I think I also
want a new app controller, that takes both the other controllers and
mediates interactions that require both of them (for instance the
current behaviour of the new button).

For now though I have only put stuff that was already in state in these
two new controllers, splitting up everything else will come later.