Make action handling neater
I wanted to have a map of state-name-action-name tuples to handlers, but
then I couldn't benefit from type narrowing, so I think this approach
makes the most sense. Nice and clean where the logic for picking what to
do is, with the function that gets called documenting what it does, and
keeping the implementation it its own space.